% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToJsonForAgol.R
\name{ToJsonForAgol}
\alias{ToJsonForAgol}
\title{GetEpiData function
This function read the running epi data online}
\usage{
ToJsonForAgol(data)
}
\description{
GetEpiData function
This function read the running epi data online
}
